move management of TEST-GUIDE servers into ATXInstallation class
- add migration routine from previous ATXPublisher configuration
- fix serialization of custom settings
- adapted tests
