+++ src/main/resources/de/tracetronic/jenkins/plugins/ecutest/report/atx/Messages.properties
@@ -5,19 +5,4 @@
 #
 ATXBuildAction.DisplayName=ATX Reports
 ATXProjectAction.DisplayName=Latest ATX Reports
-ATXPublisher.CustomAttributeExpression=Custom package attributes cannot be validated.
 ATXPublisher.DisplayName=[TT] Publish ATX Reports
-ATXPublisher.DuplicateSetting=Setting is already existing and will be ignored.
-ATXPublisher.InvalidFileExpression=Invalid file search expression.
-ATXPublisher.InvalidPort=Choose a port greater than 0 and less than 65535.
-ATXPublisher.InvalidServer=Server at {0} does not appear to be a TEST-GUIDE server.
-ATXPublisher.InvalidServerContextPath=Invalid server context path.
-ATXPublisher.InvalidServerUrl={0} is not a valid server URL.
-ATXPublisher.InvalidServerUrlProtocol=Server URL without protocol specification is expected.
-ATXPublisher.InvalidSettingName=Setting name must contain only letters.
-ATXPublisher.NeedsAdmin=To use this port on an Unix-based system root privileges are required.
-ATXPublisher.NoValidatedConnection=Parameterized server settings cannot be checked, be sure to allocate with valid \
-        values.
-ATXPublisher.NoValidatedValue=Value cannot be resolved at validation-time, be sure to allocate with a valid value.
-ATXPublisher.ServerNotReachable=Server at {0} could not be reached.\n{1}
-ATXPublisher.ValidConnection=Successfully connected to {0}
