+++ src/main/resources/de/tracetronic/jenkins/plugins/ecutest/report/atx/Messages_de.properties
@@ -5,21 +5,4 @@
 #
 ATXBuildAction.DisplayName=ATX-Reports
 ATXProjectAction.DisplayName=Letzte ATX-Reports
-ATXPublisher.CustomAttributeExpression=Benutzerdefinierte Package-Attribute k\u00f6nnen nicht \u00fcberpr\u00fcft \
-        werden.
 ATXPublisher.DisplayName=[TT] ATX-Reports ver\u00f6ffentlichen
-ATXPublisher.DuplicateSetting=Einstellung existiert bereits und wird ignoriert.
-ATXPublisher.InvalidFileExpression=Ung\u00fcltiger Datei-Suchausdruck.
-ATXPublisher.InvalidPort=Port gr\u00f6\u00dfer als 0 und kleiner als 65535 w\u00e4hlen.
-ATXPublisher.InvalidServer=Server unter {0} scheint kein TEST-GUIDE-Server zu sein.
-ATXPublisher.InvalidServerContextPath=Ung\u00fcltiges Serververzeichnis.
-ATXPublisher.InvalidServerUrl={0} ist keine g\u00fcltige Server-URL.
-ATXPublisher.InvalidServerUrlProtocol=Server-URL ohne Protokollangabe erwartet.
-ATXPublisher.InvalidSettingName=Name der Einstellung darf nur Buchstaben enthalten.
-ATXPublisher.NeedsAdmin=Um diesen Port zu verwenden, werden auf Unix-Systemen Root-Berechtigungen vorausgesetzt.
-ATXPublisher.NoValidatedConnection=Parametrisierte Server-Einstellungen k\u00f6nnen nicht direkt \u00fcberpr\u00fcft \
-        werden, Variablenzuweisungen m\u00fcssen g\u00fcltig sein.
-ATXPublisher.NoValidatedValue=Wert kann nicht direkt \u00fcberpr\u00fcft werden, Variablenzuweisung muss g\u00fcltig \
-        sein.
-ATXPublisher.ServerNotReachable=Server unter {0} kann nicht erreicht werden.\n{1}
-ATXPublisher.ValidConnection=Verbindung zu {0} erfolgreich hergestellt.
