+++ src/main/resources/de/tracetronic/jenkins/plugins/ecutest/report/atx/installation/ATXInstallation/global.properties
