+++ src/main/resources/de/tracetronic/jenkins/plugins/ecutest/report/atx/installation/Messages.properties
@@ -5,3 +5,18 @@
 #
 ATXCustomBooleanSetting.DisplayName=Boolean Setting
 ATXCustomTextSetting.DisplayName=Text Parameter Setting
+ATXInstallation.CustomAttributeExpression=Custom package attributes cannot be validated.
+ATXInstallation.DisplayName=TEST-GUIDE Installation
+ATXInstallation.DuplicateSetting=Setting is already existing and will be ignored.
+ATXInstallation.InvalidFileExpression=Invalid file search expression.
+ATXInstallation.InvalidPort=Choose a port greater than 0 and less than 65535.
+ATXInstallation.InvalidServerContextPath=Invalid server context path.
+ATXInstallation.InvalidServerUrl={0} is not a valid server URL.
+ATXInstallation.InvalidServerUrlProtocol=Server URL without protocol specification is expected.
+ATXInstallation.InvalidSettingName=Setting name must contain only letters.
+ATXInstallation.InvalidServer=Server at {0} does not appear to be a TEST-GUIDE server.
+ATXInstallation.NeedsAdmin=To use this port on an Unix-based system root privileges are required.
+ATXInstallation.NoValidatedConnection=Parameterized server settings cannot be checked, be sure to allocate with valid ones.
+ATXInstallation.NoValidatedValue=Value cannot be resolved at validation-time, be sure to allocate with a valid value.
+ATXInstallation.ServerNotReachable=Server at {0} could not be reached.\n{1}
+ATXInstallation.ValidConnection=Successfully connected to {0}
