+++ src/main/resources/de/tracetronic/jenkins/plugins/ecutest/report/atx/installation/Messages_de.properties
@@ -5,3 +5,21 @@
 #
 ATXCustomBooleanSetting.DisplayName=Boolesche Einstellung
 ATXCustomTextSetting.DisplayName=Textparameter-Einstellung
+ATXInstallation.CustomAttributeExpression=Benutzerdefinierte Package-Attribute k\u00f6nnen nicht \u00fcberpr\u00fcft \
+        werden.
+ATXInstallation.DisplayName=TEST-GUIDE-Installation
+ATXInstallation.DuplicateSetting=Einstellung existiert bereits und wird ignoriert.
+ATXInstallation.InvalidFileExpression=Ung\u00fcltiger Datei-Suchausdruck.
+ATXInstallation.InvalidPort=Port gr\u00f6\u00dfer als 0 und kleiner als 65535 w\u00e4hlen.
+ATXInstallation.InvalidServer=Server unter {0} scheint kein TEST-GUIDE-Server zu sein.
+ATXInstallation.InvalidServerContextPath=Ung\u00fcltiges Serververzeichnis.
+ATXInstallation.InvalidServerUrl={0} ist keine g\u00fcltige Server-URL.
+ATXInstallation.InvalidServerUrlProtocol=Server-URL ohne Protokollangabe erwartet.
+ATXInstallation.InvalidSettingName=Name der Einstellung darf nur Buchstaben enthalten.
+ATXInstallation.NeedsAdmin=Um diesen Port zu verwenden, werden auf Unix-Systemen Root-Berechtigungen vorausgesetzt.
+ATXInstallation.NoValidatedConnection=Parametrisierte Server-Einstellungen k\u00f6nnen nicht direkt \u00fcberpr\u00fcft \
+        werden, Variablenzuweisungen m\u00fcssen g\u00fcltig sein.
+ATXInstallation.NoValidatedValue=Wert kann nicht direkt \u00fcberpr\u00fcft werden, Variablenzuweisung muss g\u00fcltig \
+        sein.
+ATXInstallation.ServerNotReachable=Server unter {0} kann nicht erreicht werden.\n{1}
+ATXInstallation.ValidConnection=Verbindung zu {0} erfolgreich hergestellt.
