+++ src/main/resources/de/tracetronic/jenkins/plugins/ecutest/tool/installation/Messages.properties
@@ -3,13 +3,13 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 #
-ET.DisplayName=ECU-TEST
-ET.IsUnixSystem=Entry is not required because Jenkins master is running on an Unix-based system, configure each \
+ETInstallation.DisplayName=ECU-TEST
+ETInstallation.IsUnixSystem=Entry is not required because Jenkins master is running on an Unix-based system, configure each \
         individual Windows slave instead.
-ET.NoInstallation=The selected ECU-TEST installation is not configured for this node.
-ET.NotADirectory={0} is not a valid directory.
-ET.NotHomeDirectory={0} is not a valid ECU-TEST home directory.
-ET.NotRequired=Entry is mandatory only if it is intended to execute ECU-TEST on the Jenkins master, otherwise \
+ETInstallation.NoInstallation=The selected ECU-TEST installation is not configured for this node.
+ETInstallation.NotADirectory={0} is not a valid directory.
+ETInstallation.NotHomeDirectory={0} is not a valid ECU-TEST home directory.
+ETInstallation.NotRequired=Entry is mandatory only if it is intended to execute ECU-TEST on the Jenkins master, otherwise \
         configure each individual Windows slave.
 ETToolProperty.DisplayName=Advanced Settings
 ETToolProperty.InvalidProgID={0} is not a valid ProgID, allowed values are ECU-TEST.Application, \
