+++ src/test/java/de/tracetronic/jenkins/plugins/ecutest/extension/jobdsl/ReportPublisherParamInstallDslExtensionIT.java
@@ -56,8 +56,8 @@ public class ReportPublisherParamInstallDslExtensionIT extends AbstractDslExtens
             .getDescriptorByType(ETInstallation.DescriptorImpl.class);
         etDescriptor.setInstallations(new ETInstallation("ECU-TEST", "C:\\ECU-TEST", JenkinsRule.NO_PROPERTIES));
 
-        final ATXPublisher.DescriptorImpl atxImpl = jenkins.jenkins
-            .getDescriptorByType(ATXPublisher.DescriptorImpl.class);
+        final ATXInstallation.DescriptorImpl atxImpl = jenkins.jenkins
+            .getDescriptorByType(ATXInstallation.DescriptorImpl.class);
         final ATXInstallation inst = new ATXInstallation("TEST-GUIDE", "${ECUTEST}", new ATXConfig());
         atxImpl.setInstallations(inst);
     }
